Add new stations and fix URL
New site format as of today.  Haven't tested it entirely because his site's being unintentionally DOS'd by the traffic atm, so I don't want to contribute to that - but it seems to work, and I've successfully retrieved a few songs just fine given the site's new structure as of today.  I'll re-rip it all once the traffic dies down and I'll update here if I've messed anything up.