Add proper URL encoding
Previously had issues with certain track names. As you suspected in your comment on PR #4, it was indeed an encoding issue (:

Fixes the issue where track names which contain #, which is not sanitized because it's technically valid within URLs in some cases, as the prefix for a fragment identifier, like when you link someone to an anchored section of a long Wikipedia article - which... just isn't what we want here at all.